refactor(frontend): extract invitation token hook
